modify: move `config.py` into `dev` folder
