+++ ui/calc/CalcCharacter.py
@@ -5,7 +5,7 @@ from discord import ui, utils
 
 import ambr.models as ambr_models
 import dev.asset as asset
-import config
+import dev.config as config
 import dev.models as models
 from ambr import AmbrTopAPI
 from apps.db import get_user_lang, get_user_theme
