+++ ui/genshin/Build.py
@@ -6,7 +6,7 @@ import yaml
 from discord import ui
 
 import dev.asset as asset
-import config
+import dev.config as config
 from apps.db import get_user_lang, get_user_theme
 from apps.draw import main_funcs
 from apps.draw.utility import image_gen_transition
