+++ ui/genshin/EventTypeChooser.py
@@ -8,7 +8,7 @@ from discord import ui
 from discord.utils import format_dt
 
 import dev.asset as asset
-import config
+import dev.config as config
 from apps.db import get_user_lang
 from apps.hoyolab_rss_feeds.create_feed import create_feed
 from apps.text_map import text_map, to_genshin_py
