+++ ui/genshin/Leaderboard.py
@@ -4,7 +4,7 @@ import discord
 from discord import ui, utils
 
 import dev.asset as asset
-import config
+import dev.config as config
 import dev.models as models
 from ambr import AmbrTopAPI, Character
 from apps.db import get_user_theme
