+++ ui/genshin/WeaponNotificationMenu.py
@@ -4,7 +4,7 @@ import asyncpg
 from discord import Interaction, Locale, SelectOption
 from discord.ui import Button, Select
 
-import config
+import dev.config as config
 from ambr import AmbrTopAPI
 from apps.genshin import get_weapon_emoji
 from apps.text_map import text_map, to_ambr_top
