+++ ui/others/Feedback.py
@@ -1,7 +1,7 @@
 from discord import Interaction, Locale, TextStyle
 from discord.ui import TextInput
 
-import config
+import dev.config as config
 from apps.text_map import text_map
 from dev.base_ui import BaseModal
 from dev.models import DefaultEmbed
