+++ ui/others/ManageAccounts.py
@@ -8,7 +8,7 @@ from logingateway import HuTaoLoginAPI
 from logingateway.exception import UserTokenNotFound
 
 import dev.asset as asset
-import config
+import dev.config as config
 from apps.db import get_user_lang
 from apps.genshin import get_account_select_options, get_uid_region_hash
 from apps.text_map import text_map, to_hutao_login_lang
