+++ ui/others/SettingsMenu.py
@@ -4,7 +4,7 @@ import discord
 from discord import ui
 
 import dev.asset as asset
-import config
+import dev.config as config
 from apps.db import get_user_auto_redeem, get_user_lang, get_user_theme
 from apps.text_map import text_map
 from dev.base_ui import BaseView, GoBackButton
