+++ ui/wish/ChooseWeapon.py
@@ -3,7 +3,7 @@ from typing import Optional
 from discord import ButtonStyle, Interaction, Locale
 from discord.ui import Button
 
-import config
+import dev.config as config
 from apps.text_map import text_map
 from dev.base_ui import BaseView
